simplify logic for moves
Castling was a weird 1.5/2-step process. How it used to work:

The `DoMove` component is put on the king piece. A system with an
`Added<DoMove>` query filter is watching for it. It will update the
board state (from the chess crate), so the castle happens all at once.
Then the piece locations are updated; but this can only be done for the
king in the current frame. The rook entity has to be pulled from the
board state and done in the next frame by the same system. A flag is
used to ensure the board state isn't double updated.

The flag made the logic messy so it was removed. Now, moves are done in
two phases. The board state is updated in the `Added<DoMove>` system.
Then one-shot `DoUpdatePieceSquare` components are inserted onto the
pieces which get picked up by an `Added<DoUpdatePieceSquare>` system
that updates the locations.