remove redundant info in `BoardState`
`BoardState` has a map of UI piece info that contains `BoardPiece`
values. `BoardPiece` contains the entity, color, and type of a UI piece.
This information is also stored in the `chess::Board` that's in
`BoardState` and thus it is reduntant.

This commit reduces the map down to just entities and deletes the
`BoardPiece` struct.