prevent panics when entities don't exist
Use `get_*` methods on `Query` and let-else statements to return from
systems rather than using the panicking variants of the methods. This is
useful, e.g., when debugging UI because it allows for commenting out
parts of the UI without needing to disable all of the systems that used
to panic if those entities don't exist.