fix debug names
The `debug_name` macro became broken, making the format literal the name
of the entity rather than formatting it with `format!`.

Debug names with format strings are impossible to distinguish from those
that are static strings that require no formatting. Because of this,
another macro has been introduced -- `debug_name_f` -- which serves as
the "formatting-variant". Now, `debug_name` only accepts expressions
that can be forwarded to the `Name` component and string literals that
should not be formatted. The new `debug_name_f` macro only accepts valid
`format!` syntax as it will pass all of its arguments through `format!`.